* minor type conversion updates

+++ calculation/cluster/Cluster.cpp
@@ -493,7 +493,7 @@ void CCluster::DisplayNullOccurrence(FILE* fp, const CSaTScanData& Data, unsigne
   try {
     if (Data.GetParameters().GetIsProspectiveAnalysis() && Data.GetParameters().GetNumReplicationsRequested() > 98) {
       PrintFormat.PrintSectionLabel(fp, "Recurrence interval", false, true);
-      fIntervals = Data.GetNumTimeIntervals() - Data.GetProspectiveStartIndex() + 1;
+      fIntervals = static_cast<float>(Data.GetNumTimeIntervals() - Data.GetProspectiveStartIndex() + 1);
       fAdjustedP_Value = static_cast<float>(1 - pow(static_cast<double>(1 - GetPValue(iNumSimulations)), static_cast<int>(1/fIntervals)));
       fUnitsInOccurrence = (float)Data.GetParameters().GetTimeAggregationLength()/fAdjustedP_Value;
       switch (Data.GetParameters().GetTimeAggregationUnitsType()) {
